requirements.txt
weaviate-client>=3.23.1
python-dotenv>=1.0.0
openai>=0.27.9
black>=23.7.0
wasabi>=1.1.2
nltk>=3.8.1
fastapi>=0.102.0
uvicorn[standard]
click>= 8.1.7
setuptools
wheel
twine
spacy
